pyproject.toml
[tool.poetry]
name = "emu"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["Your Name <your@email.tld>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^2.2.0"
puya = "^0"
pydantic = "^2.6.1"
sqlmodel = "^0.0.16"
rich = "^13.7.0"
typer = "^0.9.0"
python-dotenv-vault = "^0.6.4"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "*"}
ruff = "^0.1.6"
mypy = "*"
pytest = "*"
pytest-cov = "*"
pip-audit = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ["E", "F", "ANN", "UP", "N", "C4", "B", "A", "YTT", "W", "FBT", "Q", "RUF", "I"]
ignore = [
  "ANN101", # no type for self
  "ANN102", # no type for cls
]
unfixable = ["B", "RUF"]

[tool.ruff.flake8-annotations]
allow-star-arg-any = true
suppress-none-returning = true

[tool.pytest.ini_options]
pythonpath = ["smart_contracts", "tests", "emu"]

[tool.mypy]
files = "smart_contracts/"
python_version = "3.12"
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
allow_untyped_defs = false
strict_equality = true
